Module aws_sdk_directconnect::model

Modules

associated_gateway

See AssociatedGateway

bgp_peer

See BgpPeer

connection

See Connection

direct_connect_gateway

See DirectConnectGateway

direct_connect_gateway_association

See DirectConnectGatewayAssociation

direct_connect_gateway_association_proposal

See DirectConnectGatewayAssociationProposal

direct_connect_gateway_attachment

See DirectConnectGatewayAttachment

interconnect

See Interconnect

lag

See Lag

loa

See Loa

location

See Location

mac_sec_key

See MacSecKey

new_bgp_peer

See NewBgpPeer

new_private_virtual_interface

See NewPrivateVirtualInterface

new_private_virtual_interface_allocation

See NewPrivateVirtualInterfaceAllocation

new_public_virtual_interface

See NewPublicVirtualInterface

new_public_virtual_interface_allocation

See NewPublicVirtualInterfaceAllocation

new_transit_virtual_interface

See NewTransitVirtualInterface

new_transit_virtual_interface_allocation

See NewTransitVirtualInterfaceAllocation

resource_tag

See ResourceTag

route_filter_prefix

See RouteFilterPrefix

tag

See Tag

virtual_gateway

See VirtualGateway

virtual_interface

See VirtualInterface

virtual_interface_test_history

See VirtualInterfaceTestHistory

Structs

AssociatedGateway

Information about the associated gateway.

BgpPeer

Information about a BGP peer.

Connection

Information about an Direct Connect connection.

DirectConnectGateway

Information about a Direct Connect gateway, which enables you to connect virtual interfaces and virtual private gateway or transit gateways.

DirectConnectGatewayAssociation

Information about an association between a Direct Connect gateway and a virtual private gateway or transit gateway.

DirectConnectGatewayAssociationProposal

Information about the proposal request to attach a virtual private gateway to a Direct Connect gateway.

DirectConnectGatewayAttachment

Information about an attachment between a Direct Connect gateway and a virtual interface.

Interconnect

Information about an interconnect.

Lag

Information about a link aggregation group (LAG).

Loa

Information about a Letter of Authorization - Connecting Facility Assignment (LOA-CFA) for a connection.

Location

Information about an Direct Connect location.

MacSecKey

Information about the MAC Security (MACsec) secret key.

NewBgpPeer

Information about a new BGP peer.

NewPrivateVirtualInterface

Information about a private virtual interface.

NewPrivateVirtualInterfaceAllocation

Information about a private virtual interface to be provisioned on a connection.

NewPublicVirtualInterface

Information about a public virtual interface.

NewPublicVirtualInterfaceAllocation

Information about a public virtual interface to be provisioned on a connection.

NewTransitVirtualInterface

Information about a transit virtual interface.

NewTransitVirtualInterfaceAllocation

Information about a transit virtual interface to be provisioned on a connection.

ResourceTag

Information about a tag associated with an Direct Connect resource.

RouteFilterPrefix

Information about a route filter prefix that a customer can advertise through Border Gateway Protocol (BGP)

Tag

Information about a tag.

VirtualGateway

Information about a virtual private gateway for a private virtual interface.

VirtualInterface

Information about a virtual interface.

VirtualInterfaceTestHistory

Information about the virtual interface failover test.

Enums

AddressFamily

BgpPeerState

BgpStatus

NOTE: BgpStatus::Unknown has been renamed to ::UnknownValue.

ConnectionState

NOTE: ConnectionState::Unknown has been renamed to ::UnknownValue.

DirectConnectGatewayAssociationProposalState

DirectConnectGatewayAssociationState

DirectConnectGatewayAttachmentState

DirectConnectGatewayAttachmentType

DirectConnectGatewayState

GatewayType

HasLogicalRedundancy

NOTE: HasLogicalRedundancy::Unknown has been renamed to ::UnknownValue.

InterconnectState

NOTE: InterconnectState::Unknown has been renamed to ::UnknownValue.

LagState

NOTE: LagState::Unknown has been renamed to ::UnknownValue.

LoaContentType

VirtualInterfaceState

NOTE: VirtualInterfaceState::Unknown has been renamed to ::UnknownValue.