Module aws_sdk_directconnect::operation

Structs

AcceptDirectConnectGatewayAssociationProposal

Accepts a proposal request to attach a virtual private gateway or transit gateway to a Direct Connect gateway.

AllocateConnectionOnInterconnect

Deprecated. Use AllocateHostedConnection instead.

AllocateHostedConnection

Creates a hosted connection on the specified interconnect or a link aggregation group (LAG) of interconnects.

AllocatePrivateVirtualInterface

Provisions a private virtual interface to be owned by the specified account.

AllocatePublicVirtualInterface

Provisions a public virtual interface to be owned by the specified account.

AllocateTransitVirtualInterface

Provisions a transit virtual interface to be owned by the specified account. Use this type of interface to connect a transit gateway to your Direct Connect gateway.

AssociateConnectionWithLag

Associates an existing connection with a link aggregation group (LAG). The connection

AssociateHostedConnection

Associates a hosted connection and its virtual interfaces with a link aggregation

AssociateMacSecKey

Associates a MAC Security (MACsec) Connection Key Name (CKN)/ Connectivity Association Key (CAK) pair with an Direct Connect dedicated connection.

AssociateVirtualInterface

Associates a virtual interface with a specified link aggregation group (LAG) or

ConfirmConnection

Confirms the creation of the specified hosted connection on an interconnect.

ConfirmPrivateVirtualInterface

Accepts ownership of a private virtual interface created by another account.

ConfirmPublicVirtualInterface

Accepts ownership of a public virtual interface created by another account.

ConfirmTransitVirtualInterface

Accepts ownership of a transit virtual interface created by another account.

CreateBGPPeer

Creates a BGP peer on the specified virtual interface.

CreateConnection

Creates a connection between a customer network and a specific Direct Connect location.

CreateDirectConnectGateway

Creates a Direct Connect gateway, which is an intermediate object that enables you to connect a set

CreateDirectConnectGatewayAssociation

Creates an association between a Direct Connect gateway and a virtual private gateway. The virtual

CreateDirectConnectGatewayAssociationProposal

Creates a proposal to associate the specified virtual private gateway or transit gateway with the specified Direct Connect gateway.

CreateInterconnect

Creates an interconnect between an Direct Connect Partner's network and a specific Direct Connect location.

CreateLag

Creates a link aggregation group (LAG) with the specified number of bundled

CreatePrivateVirtualInterface

Creates a private virtual interface. A virtual interface is the VLAN that transports Direct Connect traffic.

CreatePublicVirtualInterface

Creates a public virtual interface. A virtual interface is the VLAN that transports Direct Connect traffic.

CreateTransitVirtualInterface

Creates a transit virtual interface. A transit virtual interface should be used to access one or more transit gateways associated with Direct Connect gateways. A transit virtual interface enables the connection of multiple VPCs attached to a transit gateway to a Direct Connect gateway.

DeleteBGPPeer

Deletes the specified BGP peer on the specified virtual interface with the specified customer address and ASN.

DeleteConnection

Deletes the specified connection.

DeleteDirectConnectGateway

Deletes the specified Direct Connect gateway. You must first delete all virtual interfaces that are

DeleteDirectConnectGatewayAssociation

Deletes the association between the specified Direct Connect gateway and virtual private gateway.

DeleteDirectConnectGatewayAssociationProposal

Deletes the association proposal request between the specified Direct Connect gateway and virtual private gateway or transit gateway.

DeleteInterconnect

Deletes the specified interconnect.

DeleteLag

Deletes the specified link aggregation group (LAG). You cannot delete a LAG if it has active

DeleteVirtualInterface

Deletes a virtual interface.

DescribeConnectionLoa

Deprecated. Use DescribeLoa instead.

DescribeConnections

Displays the specified connection or all connections in this Region.

DescribeConnectionsOnInterconnect

Deprecated. Use DescribeHostedConnections instead.

DescribeDirectConnectGatewayAssociationProposals

Describes one or more association proposals for connection between a virtual private gateway or transit gateway and a Direct Connect gateway.

DescribeDirectConnectGatewayAssociations

Lists the associations between your Direct Connect gateways and virtual private gateways and transit gateways. You must specify one of the following:

DescribeDirectConnectGatewayAttachments

Lists the attachments between your Direct Connect gateways and virtual interfaces. You must specify

DescribeDirectConnectGateways

Lists all your Direct Connect gateways or only the specified Direct Connect gateway. Deleted Direct Connect gateways are not returned.

DescribeHostedConnections

Lists the hosted connections that have been provisioned on the specified

DescribeInterconnectLoa

Deprecated. Use DescribeLoa instead.

DescribeInterconnects

Lists the interconnects owned by the account or only the specified interconnect.

DescribeLags

Describes all your link aggregation groups (LAG) or the specified LAG.

DescribeLoa

Gets the LOA-CFA for a connection, interconnect, or link aggregation group (LAG).

DescribeLocations

Lists the Direct Connect locations in the current Region. These are the locations that can be selected when calling

DescribeTags

Describes the tags associated with the specified Direct Connect resources.

DescribeVirtualGateways

Lists the virtual private gateways owned by the account.

DescribeVirtualInterfaces

Displays all virtual interfaces for an account. Virtual interfaces deleted fewer

DisassociateConnectionFromLag

Disassociates a connection from a link aggregation group (LAG). The connection is

DisassociateMacSecKey

Removes the association between a MAC Security (MACsec) security key and an Direct Connect dedicated connection.

ListVirtualInterfaceTestHistory

Lists the virtual interface failover test history.

StartBgpFailoverTest

Starts the virtual interface failover test that verifies your configuration meets your resiliency requirements by placing the BGP peering session in the DOWN state. You can then send traffic to verify that there are no outages.

StopBgpFailoverTest

Stops the virtual interface failover test.

TagResource

Adds the specified tags to the specified Direct Connect resource. Each resource can have a maximum of 50 tags.

UntagResource

Removes one or more tags from the specified Direct Connect resource.

UpdateConnection

Updates the Direct Connect dedicated connection configuration.

UpdateDirectConnectGatewayAssociation

Updates the specified attributes of the Direct Connect gateway association.

UpdateLag

Updates the attributes of the specified link aggregation group (LAG).

UpdateVirtualInterfaceAttributes

Updates the specified attributes of the specified virtual private interface.