Struct aws_sdk_directconnect::output::allocate_private_virtual_interface_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for AllocatePrivateVirtualInterfaceOutput

Implementations

impl Builder

pub fn owner_account(self, input: impl Into<String>) -> Self

The ID of the account that owns the virtual interface.

pub fn set_owner_account(self, input: Option<String>) -> Self

pub fn virtual_interface_id(self, input: impl Into<String>) -> Self

The ID of the virtual interface.

pub fn set_virtual_interface_id(self, input: Option<String>) -> Self

pub fn location(self, input: impl Into<String>) -> Self

The location of the connection.

pub fn set_location(self, input: Option<String>) -> Self

pub fn connection_id(self, input: impl Into<String>) -> Self

The ID of the connection.

pub fn set_connection_id(self, input: Option<String>) -> Self

pub fn virtual_interface_type(self, input: impl Into<String>) -> Self

The type of virtual interface. The possible values are private and public.

pub fn set_virtual_interface_type(self, input: Option<String>) -> Self

pub fn virtual_interface_name(self, input: impl Into<String>) -> Self

The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

pub fn set_virtual_interface_name(self, input: Option<String>) -> Self

pub fn vlan(self, input: i32) -> Self

The ID of the VLAN.

pub fn set_vlan(self, input: Option<i32>) -> Self

pub fn asn(self, input: i32) -> Self

The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.

pub fn set_asn(self, input: Option<i32>) -> Self

pub fn amazon_side_asn(self, input: i64) -> Self

The autonomous system number (ASN) for the Amazon side of the connection.

pub fn set_amazon_side_asn(self, input: Option<i64>) -> Self

pub fn auth_key(self, input: impl Into<String>) -> Self

The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

pub fn set_auth_key(self, input: Option<String>) -> Self

pub fn amazon_address(self, input: impl Into<String>) -> Self

The IP address assigned to the Amazon interface.

pub fn set_amazon_address(self, input: Option<String>) -> Self

pub fn customer_address(self, input: impl Into<String>) -> Self

The IP address assigned to the customer interface.

pub fn set_customer_address(self, input: Option<String>) -> Self

pub fn address_family(self, input: AddressFamily) -> Self

The address family for the BGP peer.

pub fn set_address_family(self, input: Option<AddressFamily>) -> Self

pub fn virtual_interface_state(self, input: VirtualInterfaceState) -> Self

The state of the virtual interface. The following are the possible values:

confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner.

verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created.

pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic.

available: A virtual interface that is able to forward traffic.

down: A virtual interface that is BGP down.

deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic.

deleted: A virtual interface that cannot forward traffic.

rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface enters the Rejected state.

unknown: The state of the virtual interface is not available.

pub fn set_virtual_interface_state(
self,
input: Option<VirtualInterfaceState>
) -> Self

pub fn customer_router_config(self, input: impl Into<String>) -> Self

The customer router configuration.

pub fn set_customer_router_config(self, input: Option<String>) -> Self

pub fn mtu(self, input: i32) -> Self

The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 9001. The default value is 1500.

pub fn set_mtu(self, input: Option<i32>) -> Self

pub fn jumbo_frame_capable(self, input: bool) -> Self

Indicates whether jumbo frames (9001 MTU) are supported.

pub fn set_jumbo_frame_capable(self, input: Option<bool>) -> Self

pub fn virtual_gateway_id(self, input: impl Into<String>) -> Self

The ID of the virtual private gateway. Applies only to private virtual interfaces.

pub fn set_virtual_gateway_id(self, input: Option<String>) -> Self

pub fn direct_connect_gateway_id(self, input: impl Into<String>) -> Self

The ID of the Direct Connect gateway.

pub fn set_direct_connect_gateway_id(self, input: Option<String>) -> Self

pub fn route_filter_prefixes(self, input: impl Into<RouteFilterPrefix>) -> Self

pub fn set_route_filter_prefixes(
self,
input: Option<Vec<RouteFilterPrefix>>
) -> Self

pub fn bgp_peers(self, input: impl Into<BgpPeer>) -> Self

pub fn set_bgp_peers(self, input: Option<Vec<BgpPeer>>) -> Self

pub fn region(self, input: impl Into<String>) -> Self

The Region where the virtual interface is located.

pub fn set_region(self, input: Option<String>) -> Self

pub fn aws_device_v2(self, input: impl Into<String>) -> Self

The Direct Connect endpoint that terminates the physical connection.

pub fn set_aws_device_v2(self, input: Option<String>) -> Self

pub fn aws_logical_device_id(self, input: impl Into<String>) -> Self

The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

pub fn set_aws_logical_device_id(self, input: Option<String>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn build(self) -> AllocatePrivateVirtualInterfaceOutput

Consumes the builder and constructs a AllocatePrivateVirtualInterfaceOutput