Struct aws_sdk_directconnect::output::AllocatePublicVirtualInterfaceOutput

#[non_exhaustive]
pub struct AllocatePublicVirtualInterfaceOutput {

    pub owner_account: Option<String>,
    pub virtual_interface_id: Option<String>,
    pub location: Option<String>,
    pub connection_id: Option<String>,
    pub virtual_interface_type: Option<String>,
    pub virtual_interface_name: Option<String>,
    pub vlan: i32,
    pub asn: i32,
    pub amazon_side_asn: Option<i64>,
    pub auth_key: Option<String>,
    pub amazon_address: Option<String>,
    pub customer_address: Option<String>,
    pub address_family: Option<AddressFamily>,
    pub virtual_interface_state: Option<VirtualInterfaceState>,
    pub customer_router_config: Option<String>,
    pub mtu: Option<i32>,
    pub jumbo_frame_capable: Option<bool>,
    pub virtual_gateway_id: Option<String>,
    pub direct_connect_gateway_id: Option<String>,
    pub route_filter_prefixes: Option<Vec<RouteFilterPrefix>>,
    pub bgp_peers: Option<Vec<BgpPeer>>,
    pub region: Option<String>,
    pub aws_device_v2: Option<String>,
    pub aws_logical_device_id: Option<String>,
    pub tags: Option<Vec<Tag>>,

}

Information about a virtual interface.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

owner_account: Option<String>

The ID of the account that owns the virtual interface.

virtual_interface_id: Option<String>

The ID of the virtual interface.

location: Option<String>

The location of the connection.

connection_id: Option<String>

The ID of the connection.

virtual_interface_type: Option<String>

The type of virtual interface. The possible values are private and public.

virtual_interface_name: Option<String>

The name of the virtual interface assigned by the customer network. The name has a maximum of 100 characters. The following are valid characters: a-z, 0-9 and a hyphen (-).

vlan: i32

The ID of the VLAN.

asn: i32

The autonomous system (AS) number for Border Gateway Protocol (BGP) configuration.

The valid values are 1-2147483647.

amazon_side_asn: Option<i64>

The autonomous system number (ASN) for the Amazon side of the connection.

auth_key: Option<String>

The authentication key for BGP configuration. This string has a minimum length of 6 characters and and a maximun lenth of 80 characters.

amazon_address: Option<String>

The IP address assigned to the Amazon interface.

customer_address: Option<String>

The IP address assigned to the customer interface.

address_family: Option<AddressFamily>

The address family for the BGP peer.

virtual_interface_state: Option<VirtualInterfaceState>

The state of the virtual interface. The following are the possible values:

confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner.

verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created.

pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic.

available: A virtual interface that is able to forward traffic.

down: A virtual interface that is BGP down.

deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic.

deleted: A virtual interface that cannot forward traffic.

rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface enters the Rejected state.

unknown: The state of the virtual interface is not available.

customer_router_config: Option<String>

The customer router configuration.

mtu: Option<i32>

The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 9001. The default value is 1500.

jumbo_frame_capable: Option<bool>

Indicates whether jumbo frames (9001 MTU) are supported.

virtual_gateway_id: Option<String>

The ID of the virtual private gateway. Applies only to private virtual interfaces.

direct_connect_gateway_id: Option<String>

The ID of the Direct Connect gateway.

route_filter_prefixes: Option<Vec<RouteFilterPrefix>>

The routes to be advertised to the Amazon Web Services network in this Region. Applies to public virtual interfaces.

bgp_peers: Option<Vec<BgpPeer>>

The BGP peers configured on this virtual interface.

region: Option<String>

The Region where the virtual interface is located.

aws_device_v2: Option<String>

The Direct Connect endpoint that terminates the physical connection.

aws_logical_device_id: Option<String>

The Direct Connect endpoint that terminates the logical connection. This device might be different than the device that terminates the physical connection.

tags: Option<Vec<Tag>>

The tags associated with the virtual interface.

Implementations

impl AllocatePublicVirtualInterfaceOutput

pub fn builder() -> Builder

Creates a new builder-style object to manufacture AllocatePublicVirtualInterfaceOutput